# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Racks API documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/rack/{id}": {
            "get": {
                "tags": [
                    "Rack"
                ],
                "summary": "Get rack",
                "description": "Get rack",
                "operationId": "RetrieveRack",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rack ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get rack",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/RetrieveRackResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such rack",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchRackForRackResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rack"
                ],
                "summary": "Delete rack",
                "description": "Delete rack",
                "operationId": "DeleteRack",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rack ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Rack deleted - No content"
                    },
                    "404": {
                        "description": "No such rack",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchRackForRackResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/UnableToDeleteRackResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Action not allowed for this department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/PermissionExceptionResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Rack"
                ],
                "summary": "Update rack",
                "description": "Update rack",
                "operationId": "UpdateRack",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rack ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/UpdateRackRequest"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Rack updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/RackUpdatedResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such rack",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchRackForRackResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/RackNameExceptionResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Action not allowed for this department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/PermissionExceptionResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/UnableToUpdateRackResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/rack/vendors": {
            "get": {
                "tags": [
                    "Rack"
                ],
                "summary": "Get rack vendors",
                "description": "Get rack vendors list",
                "operationId": "GetRackVendors",
                "responses": {
                    "200": {
                        "description": "Get rack vendors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "properties": {
                                                        "item_type": {
                                                            "type": "string",
                                                            "example": "rack_vendor"
                                                        },
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "description": "Rack vendor",
                                                                "type": "string"
                                                            },
                                                            "example": [
                                                                "ITK",
                                                                "ZPAS"
                                                            ]
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PermissionExceptionResponse": {
                "title": "Action not allowed for this department",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Action not allowed for this department"
                    }
                },
                "type": "object"
            },
...
            "UpdateRackRequest": {
                "title": "Update rack request",
                "properties": {
                    "name": {
                        "description": "Rack name",
                        "type": "string",
                        "example": "Rack name",
                        "nullable": true
                    },
                    "vendor": {
                        "description": "Rack vendor",
                        "type": "string",
                        "example": "Rack vendor",
                        "nullable": true
                    },
                    "model": {
                        "description": "Rack model",
                        "type": "string",
                        "example": "Rack model",
                        "nullable": true
                    },
                    "description": {
                        "description": "Rack info",
                        "type": "string",
                        "example": "Some info about rack",
                        "nullable": true
                    },
                    "has_numbering_from_top_to_bottom": {
                        "description": "Rack numbering direction",
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "responsible": {
                        "description": "Rack responsible",
                        "type": "string",
                        "example": "Smith W.",
                        "nullable": true
                    },
                    "financially_responsible_person": {
                        "description": "Rack financially responsible",
                        "type": "string",
                        "example": "Smith W.",
                        "nullable": true
                    },
                    "inventory_number": {
                        "description": "Rack inventory number",
                        "type": "string",
                        "example": "RTF1234563",
                        "nullable": true
                    },
                    "fixed_asset": {
                        "description": "Fixed asset",
                        "type": "string",
                        "example": "1234563",
                        "nullable": true
                    },
                    "link_to_docs": {
                        "description": "Link to documents",
                        "type": "string",
                        "example": "F:\\Docs",
                        "nullable": true
                    },
                    "row": {
                        "description": "Row",
                        "type": "string",
                        "example": "1st",
                        "nullable": true
                    },
                    "place": {
                        "description": "Place",
                        "type": "string",
                        "example": "2nd",
                        "nullable": true
                    },
                    "height": {
                        "description": "Rack height (mm)",
                        "type": "integer",
                        "example": 1500,
                        "nullable": true
                    },
                    "width": {
                        "description": "Rack width (mm)",
                        "type": "integer",
                        "example": 300,
                        "nullable": true
                    },
                    "depth": {
                        "description": "Rack depth (mm)",
                        "type": "integer",
                        "example": 300,
                        "nullable": true
                    },
                    "unit_width": {
                        "description": "Rack unit width (mm)",
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "unit_depth": {
                        "description": "Rack unit depth (mm)",
                        "type": "integer",
                        "example": 100,
                        "nullable": true
                    },
                    "type": {
                        "description": "Rack type",
                        "type": "string",
                        "enum": [
                            "Rack",
                            "Protective cabinet"
                        ],
                        "example": "Rack",
                        "nullable": true
                    },
                    "frame": {
                        "description": "Rack frame",
                        "type": "string",
                        "enum": [
                            "Single frame",
                            "Double frame"
                        ],
                        "example": "Single frame",
                        "nullable": true
                    },
                    "place_type": {
                        "description": "Rack place type",
                        "type": "string",
                        "enum": [
                            "Floor standing",
                            "Wall mounted"
                        ],
                        "example": "Floor standing",
                        "nullable": true
                    },
                    "max_load": {
                        "description": "Rack max load (kilo)",
                        "type": "integer",
                        "example": 1000,
                        "nullable": true
                    },
                    "power_sockets": {
                        "description": "Power sockets",
                        "type": "integer",
                        "example": 10,
                        "nullable": true
                    },
                    "power_sockets_ups": {
                        "description": "Power sockets UPS",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "has_external_ups": {
                        "description": "Has external UPS",
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    },
                    "has_cooler": {
                        "description": "Has cooler",
                        "type": "boolean",
                        "example": false,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "NoSuchRackForRackResponse": {
                "title": "No such rack",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No such rack"
                    }
                },
                "type": "object"
            },
...
            "RackNameExceptionResponse": {
                "title": "Rack with this name already exists in this room",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Rack with this name already exists in this room"
                    }
                },
                "type": "object"
            },
...
            "RetrieveRackResponse": {
                "title": "Get rack",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Rack name"
                    },
                    "room_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "amount": {
                        "type": "integer",
                        "example": 20
                    },
                    "busy_units": {
                        "properties": {
                            "front": {
                                "type": "array",
                                "items": {
                                    "description": "Unit number",
                                    "type": "number"
                                }
                            },
                            "back": {
                                "type": "array",
                                "items": {
                                    "description": "Unit number",
                                    "type": "number"
                                }
                            }
                        },
                        "type": "object",
                        "example": {
                            "front": [
                                1,
                                2,
                                3
                            ],
                            "back": [
                                10,
                                11,
                                12
                            ]
                        }
                    },
                    "vendor": {
                        "type": "string",
                        "example": "Rack vendor"
                    },
                    "model": {
                        "type": "string",
                        "example": "Rack model"
                    },
                    "description": {
                        "type": "string",
                        "example": "Some info about rack"
                    },
                    "has_numbering_from_top_to_bottom": {
                        "type": "boolean",
                        "example": false
                    },
                    "responsible": {
                        "type": "string",
                        "example": "Smith W."
                    },
                    "financially_responsible_person": {
                        "type": "string",
                        "example": "Smith W."
                    },
                    "inventory_number": {
                        "type": "string",
                        "example": "RTF1234563"
                    },
                    "fixed_asset": {
                        "type": "string",
                        "example": "1234563"
                    },
                    "link_to_docs": {
                        "type": "string",
                        "example": "F:\\Docs"
                    },
                    "row": {
                        "type": "string",
                        "example": "1st"
                    },
                    "place": {
                        "type": "string",
                        "example": "2nd"
                    },
                    "height": {
                        "type": "integer",
                        "example": 1500
                    },
                    "width": {
                        "type": "integer",
                        "example": 300
                    },
                    "depth": {
                        "type": "integer",
                        "example": 300
                    },
                    "unit_width": {
                        "type": "integer",
                        "example": 100
                    },
                    "unit_depth": {
                        "type": "integer",
                        "example": 100
                    },
                    "type": {
                        "description": "Rack type",
                        "type": "string",
                        "enum": [
                            "Rack",
                            "Protective cabinet"
                        ],
                        "example": "Rack"
                    },
                    "frame": {
                        "description": "Rack frame",
                        "type": "string",
                        "enum": [
                            "Single frame",
                            "Double frame"
                        ],
                        "example": "Single frame"
                    },
                    "place_type": {
                        "description": "Rack place type",
                        "type": "string",
                        "enum": [
                            "Floor standing",
                            "Wall mounted"
                        ],
                        "example": "Floor standing"
                    },
                    "max_load": {
                        "type": "integer",
                        "example": 1000
                    },
                    "power_sockets": {
                        "type": "integer",
                        "example": 10
                    },
                    "power_sockets_ups": {
                        "type": "integer",
                        "example": 5
                    },
                    "has_external_ups": {
                        "type": "boolean",
                        "example": false
                    },
                    "has_cooler": {
                        "type": "boolean",
                        "example": false
                    },
                    "updated_by": {
                        "type": "string",
                        "example": "Username"
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    }
                },
                "type": "object"
            },
...
            "UnableToDeleteRackResponse": {
                "title": "Unable to delete rack",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unable to delete rack"
                    },
                    "error": {
                        "type": "string",
                        "example": "Some internal error"
                    }
                },
                "type": "object"
            },
            "UnableToUpdateRackResponse": {
                "title": "Unable to update rack",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unable to update rack"
                    },
                    "error": {
                        "type": "string",
                        "example": "Some internal error"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}